refactor(tauri)!: delete legacy profiles config module
