+++ backend/tauri/src/config/mod.rs
@@ -2,14 +2,7 @@ mod clash;
 mod core;
 mod draft;
 pub mod nyanpasu;
-pub mod profile;
 mod runtime;
-pub use self::{
-    clash::*,
-    core::*,
-    draft::*,
-    profile::{item::*, profiles::*},
-    runtime::*,
-};
+pub use self::{clash::*, core::*, draft::*, runtime::*};
 
 pub use self::nyanpasu::IVerge;
