--- backend/tauri/src/config/profile/item/prelude.rs
@@ -1,4 +1,0 @@
-#![allow(unused_imports)]
-pub use super::{
-    ProfileCleanup, ProfileFileIo, ProfileHelper, ProfileKindGetter, ProfileMetaGetter,
-};
