--- backend/tauri/src/config/profile/mod.rs
@@ -1,10 +1,0 @@
-pub mod builder;
-pub mod item;
-pub mod item_type;
-pub mod profiles;
-
-pub use builder::ProfileBuilder;
-use item::deserialize_single_or_vec;
-
-#[cfg(test)]
-mod tests;
